requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Jinja2
Werkzeug
pyjwt
validators
